Update create_mofa.R
For datasets that have a modality (and groups information) with N cells and another modality that has less than N cells, using the groups metadata for the second modality will result in indexing outside of the range of the smaller modality and throw an error when generating the MOFA object. This ensures that the underrepresented modality is not indexed out of its range.